+++ src/App.css
@@ -1,0 +1,5 @@
+.app{
+    display: grid;
+    grid-template-columns: 16% 80%;
+    column-gap: 4%;
+}
