--- src/Layout/header/index.tsx
@@ -1,5 +1,0 @@
-const Header = (): JSX.Element => {
-    return (
-        <></>
-    )
-}
