+++ src/Layout/sider/index.css
@@ -1,0 +1,6 @@
+.navigation-bar{
+    /* position: fixed; */
+    height: 100vh;
+    background-color: aqua;
+
+}
