+++ src/index.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { BrowserRouter as Router, Routes, Route, Link,  } from 'react-router-dom';
 import './index.css';
-import Main from './Layout/main/main';
+import App from './App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 
 root.render(
-    <Main />
-);
+  <Router>
+    <App/>
+  </Router>,
+)
